Rename ttl-mode.el.in to ttl-mode.el and hard-code version string
The old code in the Makefile that copied "ttl-mode.el.in" to
"ttl-mode.el" and then replaced the placeholder @VERSION@ in that copy
with the version taken from the RELEASE variable set in the Makefile,
does not work anymore because it relied on Mercurial being used but
this project now lives in a Git repository.

This machinery did not actually save the maintainer from having to edit
a file to bump the version, it was just a different file.  So there was
no benefit to it.  But there were drawbacks: users had to generate the
real "ttl-mode.el", and this is one out of only two packages on Melpa,
that made it necessary to implement a kludge to rename a "*.el.in" file
to "*.el".